Pull master

+++ homeassistant/components/recorder/migration.py
@@ -1,19 +1,24 @@
 """Schema migration helpers."""
 import logging
-import os
 
+from sqlalchemy import ForeignKeyConstraint, MetaData, Table, text
+from sqlalchemy.engine import reflection
+from sqlalchemy.exc import (
+    InternalError,
+    OperationalError,
+    ProgrammingError,
+    SQLAlchemyError,
+)
+from sqlalchemy.schema import AddConstraint, DropConstraint
+
+from .models import SCHEMA_VERSION, TABLE_STATES, Base, SchemaChanges
 from .util import session_scope
 
 _LOGGER = logging.getLogger(__name__)
-PROGRESS_FILE = ".migration_progress"
 
 
-def migrate_schema(instance):
-    """Check if the schema needs to be upgraded."""
-    from .models import SchemaChanges, SCHEMA_VERSION
-
-    progress_path = instance.hass.config.path(PROGRESS_FILE)
-
+def get_schema_version(instance):
+    """Get the schema version."""
     with session_scope(session=instance.get_session()) as session:
         res = (
             session.query(SchemaChanges)
@@ -28,30 +33,27 @@ def migrate_schema(instance):
                 "No schema version found. Inspected version: %s", current_version
             )
 
-        if current_version == SCHEMA_VERSION:
-            # Clean up if old migration left file
-            if os.path.isfile(progress_path):
-                _LOGGER.warning("Found existing migration file, cleaning up")
-                os.remove(instance.hass.config.path(PROGRESS_FILE))
-            return
+        return current_version
 
-        with open(progress_path, "w"):
-            pass
 
+def schema_is_current(current_version):
+    """Check if the schema is current."""
+    return current_version == SCHEMA_VERSION
+
+
+def migrate_schema(instance, current_version):
+    """Check if the schema needs to be upgraded."""
+    with session_scope(session=instance.get_session()) as session:
         _LOGGER.warning(
             "Database is about to upgrade. Schema version: %s", current_version
         )
+        for version in range(current_version, SCHEMA_VERSION):
+            new_version = version + 1
+            _LOGGER.info("Upgrading recorder db schema to version %s", new_version)
+            _apply_update(instance.engine, new_version, current_version)
+            session.add(SchemaChanges(schema_version=new_version))
 
-        try:
-            for version in range(current_version, SCHEMA_VERSION):
-                new_version = version + 1
-                _LOGGER.info("Upgrading recorder db schema to version %s", new_version)
-                _apply_update(instance.engine, new_version, current_version)
-                session.add(SchemaChanges(schema_version=new_version))
-
-                _LOGGER.info("Upgrade to version %s done", new_version)
-        finally:
-            os.remove(instance.hass.config.path(PROGRESS_FILE))
+            _LOGGER.info("Upgrade to version %s done", new_version)
 
 
 def _create_index(engine, table_name, index_name):
@@ -60,16 +62,16 @@ def _create_index(engine, table_name, index_name):
     The index name should match the name given for the index
     within the table definition described in the models
     """
-    from sqlalchemy import Table
-    from sqlalchemy.exc import OperationalError
-    from . import models
-
-    table = Table(table_name, models.Base.metadata)
-    _LOGGER.debug("Looking up index for table %s", table_name)
+    table = Table(table_name, Base.metadata)
+    _LOGGER.debug("Looking up index %s for table %s", index_name, table_name)
     # Look up the index object by name from the table is the models
-    index = next(idx for idx in table.indexes if idx.name == index_name)
+    index_list = [idx for idx in table.indexes if idx.name == index_name]
+    if not index_list:
+        _LOGGER.debug("The index %s no longer exists", index_name)
+        return
+    index = index_list[0]
     _LOGGER.debug("Creating %s index", index_name)
-    _LOGGER.info(
+    _LOGGER.warning(
         "Adding index `%s` to database. Note: this can take several "
         "minutes on large databases and slow computers. Please "
         "be patient!",
@@ -77,8 +79,10 @@ def _create_index(engine, table_name, index_name):
     )
     try:
         index.create(engine)
-    except OperationalError as err:
-        if "already exists" not in str(err).lower():
+    except (InternalError, ProgrammingError, OperationalError) as err:
+        lower_err_str = str(err).lower()
+
+        if "already exists" not in lower_err_str and "duplicate" not in lower_err_str:
             raise
 
         _LOGGER.warning(
@@ -99,9 +103,6 @@ def _drop_index(engine, table_name, index_name):
     string here is generated from the method parameters without sanitizing.
     DO NOT USE THIS FUNCTION IN ANY OPERATION THAT TAKES USER INPUT.
     """
-    from sqlalchemy import text
-    from sqlalchemy.exc import SQLAlchemyError
-
     _LOGGER.debug("Dropping index %s from table %s", index_name, table_name)
     success = False
 
@@ -148,10 +149,15 @@ def _drop_index(engine, table_name, index_name):
             "Finished dropping index %s from table %s", index_name, table_name
         )
     else:
+        if index_name == "ix_states_context_parent_id":
+            # Was only there on nightly so we do not want
+            # to generate log noise or issues about it.
+            return
+
         _LOGGER.warning(
             "Failed to drop index %s from table %s. Schema "
             "Migration will continue; this is not a "
-            "critical operation.",
+            "critical operation",
             index_name,
             table_name,
         )
@@ -159,10 +165,7 @@ def _drop_index(engine, table_name, index_name):
 
 def _add_columns(engine, table_name, columns_def):
     """Add columns to a table."""
-    from sqlalchemy import text
-    from sqlalchemy.exc import OperationalError
-
-    _LOGGER.info(
+    _LOGGER.warning(
         "Adding columns %s to table %s. Note: this can take several "
         "minutes on large databases and slow computers. Please "
         "be patient!",
@@ -181,10 +184,10 @@ def _add_columns(engine, table_name, columns_def):
             )
         )
         return
-    except OperationalError:
+    except (InternalError, OperationalError):
         # Some engines support adding all columns at once,
         # this error is when they don't
-        _LOGGER.info("Unable to use quick column add. Adding 1 by 1.")
+        _LOGGER.info("Unable to use quick column add. Adding 1 by 1")
 
     for column_def in columns_def:
         try:
@@ -195,7 +198,7 @@ def _add_columns(engine, table_name, columns_def):
                     )
                 )
             )
-        except OperationalError as err:
+        except (InternalError, OperationalError) as err:
             if "duplicate" not in str(err).lower():
                 raise
 
@@ -203,6 +206,104 @@ def _add_columns(engine, table_name, columns_def):
                 "Column %s already exists on %s, continuing",
                 column_def.split(" ")[1],
                 table_name,
+            )
+
+
+def _modify_columns(engine, table_name, columns_def):
+    """Modify columns in a table."""
+    if engine.dialect.name == "sqlite":
+        _LOGGER.debug(
+            "Skipping to modify columns %s in table %s; "
+            "Modifying column length in SQLite is unnecessary, "
+            "it does not impose any length restrictions",
+            ", ".join(column.split(" ")[0] for column in columns_def),
+            table_name,
+        )
+        return
+
+    _LOGGER.warning(
+        "Modifying columns %s in table %s. Note: this can take several "
+        "minutes on large databases and slow computers. Please "
+        "be patient!",
+        ", ".join(column.split(" ")[0] for column in columns_def),
+        table_name,
+    )
+
+    if engine.dialect.name == "postgresql":
+        columns_def = [
+            "ALTER {column} TYPE {type}".format(
+                **dict(zip(["column", "type"], col_def.split(" ", 1)))
+            )
+            for col_def in columns_def
+        ]
+    elif engine.dialect.name == "mssql":
+        columns_def = [f"ALTER COLUMN {col_def}" for col_def in columns_def]
+    else:
+        columns_def = [f"MODIFY {col_def}" for col_def in columns_def]
+
+    try:
+        engine.execute(
+            text(
+                "ALTER TABLE {table} {columns_def}".format(
+                    table=table_name, columns_def=", ".join(columns_def)
+                )
+            )
+        )
+        return
+    except (InternalError, OperationalError):
+        _LOGGER.info("Unable to use quick column modify. Modifying 1 by 1")
+
+    for column_def in columns_def:
+        try:
+            engine.execute(
+                text(
+                    "ALTER TABLE {table} {column_def}".format(
+                        table=table_name, column_def=column_def
+                    )
+                )
+            )
+        except (InternalError, OperationalError):
+            _LOGGER.exception(
+                "Could not modify column %s in table %s", column_def, table_name
+            )
+
+
+def _update_states_table_with_foreign_key_options(engine):
+    """Add the options to foreign key constraints."""
+    inspector = reflection.Inspector.from_engine(engine)
+    alters = []
+    for foreign_key in inspector.get_foreign_keys(TABLE_STATES):
+        if foreign_key["name"] and (
+            # MySQL/MariaDB will have empty options
+            not foreign_key["options"]
+            or
+            # Postgres will have ondelete set to None
+            foreign_key["options"].get("ondelete") is None
+        ):
+            alters.append(
+                {
+                    "old_fk": ForeignKeyConstraint((), (), name=foreign_key["name"]),
+                    "columns": foreign_key["constrained_columns"],
+                }
+            )
+
+    if not alters:
+        return
+
+    states_key_constraints = Base.metadata.tables[TABLE_STATES].foreign_key_constraints
+    old_states_table = Table(  # noqa: F841 pylint: disable=unused-variable
+        TABLE_STATES, MetaData(), *[alter["old_fk"] for alter in alters]
+    )
+
+    for alter in alters:
+        try:
+            engine.execute(DropConstraint(alter["old_fk"]))
+            for fkc in states_key_constraints:
+                if fkc.column_keys == alter["columns"]:
+                    engine.execute(AddConstraint(fkc))
+        except (InternalError, OperationalError):
+            _LOGGER.exception(
+                "Could not update foreign options in %s table", TABLE_STATES
             )
 
 
@@ -256,14 +357,54 @@ def _apply_update(engine, new_version, old_version):
     elif new_version == 7:
         _create_index(engine, "states", "ix_states_entity_id")
     elif new_version == 8:
-        # Pending migration, want to group a few.
+        _add_columns(engine, "events", ["context_parent_id CHARACTER(36)"])
+        _add_columns(engine, "states", ["old_state_id INTEGER"])
+        _create_index(engine, "events", "ix_events_context_parent_id")
+    elif new_version == 9:
+        # We now get the context from events with a join
+        # since its always there on state_changed events
+        #
+        # Ideally we would drop the columns from the states
+        # table as well but sqlite doesn't support that
+        # and we would have to move to something like
+        # sqlalchemy alembic to make that work
+        #
+        _drop_index(engine, "states", "ix_states_context_id")
+        _drop_index(engine, "states", "ix_states_context_user_id")
+        # This index won't be there if they were not running
+        # nightly but we don't treat that as a critical issue
+        _drop_index(engine, "states", "ix_states_context_parent_id")
+        # Redundant keys on composite index:
+        # We already have ix_states_entity_id_last_updated
+        _drop_index(engine, "states", "ix_states_entity_id")
+        _create_index(engine, "events", "ix_events_event_type_time_fired")
+        _drop_index(engine, "events", "ix_events_event_type")
+    elif new_version == 10:
+        # Now done in step 11
         pass
-        # _add_columns(engine, "events", [
-        #     'context_parent_id CHARACTER(36)',
-        # ])
-        # _add_columns(engine, "states", [
-        #     'context_parent_id CHARACTER(36)',
-        # ])
+    elif new_version == 11:
+        _create_index(engine, "states", "ix_states_old_state_id")
+        _update_states_table_with_foreign_key_options(engine)
+    elif new_version == 12:
+        if engine.dialect.name == "mysql":
+            _modify_columns(engine, "events", ["event_data LONGTEXT"])
+            _modify_columns(engine, "states", ["attributes LONGTEXT"])
+    elif new_version == 13:
+        if engine.dialect.name == "mysql":
+            _modify_columns(
+                engine, "events", ["time_fired DATETIME(6)", "created DATETIME(6)"]
+            )
+            _modify_columns(
+                engine,
+                "states",
+                [
+                    "last_changed DATETIME(6)",
+                    "last_updated DATETIME(6)",
+                    "created DATETIME(6)",
+                ],
+            )
+    elif new_version == 14:
+        _modify_columns(engine, "events", ["event_type VARCHAR(64)"])
     else:
         raise ValueError(f"No schema migration defined for version {new_version}")
 
@@ -277,9 +418,6 @@ def _inspect_schema_version(engine, session):
     version 1 are present to make the determination. Eventually this logic
     can be removed and we can assume a new db is being created.
     """
-    from sqlalchemy.engine import reflection
-    from .models import SchemaChanges, SCHEMA_VERSION
-
     inspector = reflection.Inspector.from_engine(engine)
     indexes = inspector.get_indexes("events")
